Add files via upload
The input is given in the path variable, which contains the full path to a .txt file that contains the list of nodes connected to form a network, e.g.  id1, id13, id 5, id9, id10 
                      id2, id6, id15, id8
                      id3, id9, id22, id14, id16
Meaning that node1 is connected to nodes 13, 5, 9 and 10, etc. 
The resulting adjacency matrix is written to AdjacencyInSparse.txt in the form of sparse matrix notation, e.g.:
 1, 13
 1, 5
 1, 9
 1, 10
 2, 6 ..........
meaning again that node1 connects to nodes 13, 5, 9 and 10, etc.